Check if there's an existing grid
I think I found a way to check if there's an existing grid before
removing it. The problem with trying to remove it so far is that I get
a reference error for an undefined variable if there are no rows or
columns. The following methods after the check don't remove the grid,
though. I have to work on that next.