Remove existing grid, if there is one
I got stuck here because I was trying to remove row and column, but they
are block-scoped inside the loop, and defined after you've tried to
reset the grid. So they don't exist at that point, and will never exist
outside the scope where you've defined them.

I knew that, but didn't know how to get around it.

The solution was that grid, the variable you're using to hold your
rows/columns/cells? It does exist. And you can check if that has a
firstChild property - and while it does, keep remove()-ing that
firstChild.